[IMP] Add context to the tasks so they can be cacelled.
This PR adds context to the tasks so when they are started can be cancelled (if they consider the context cancellation), before this PR the task can be killed only if they haven't started.

Also some minor typos.

These changen brake the api, but can easily fixed just adding the context parameter.